Update dictyionary keys.py
Hey, I noticed the current code is hardcoded to only work for numbers between 1 and 20. What if someone wants to calculate squares for a different range, like 25 to 30? The current code won’t handle that. I made some changes to make it more flexible so it can work for any range of numbers. I also added a check to make sure the start value is less than the end value, so it doesn’t produce weird results if someone accidentally passes invalid inputs.